// contracts/generated/typescript/models/WriteCollectionMember.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Anby Wiki API
 * Anby Wiki HTTP API 权威契约（OpenAPI 3.1）。 本文件是所有接口行为的唯一权威定义：错误模型、分页、幂等头与请求 ID 在此统一定义。 前端客户端由 OpenAPI Generator 从本文件生成，禁止手写 URL/DTO。 Go 侧 handler 手写实现，由 Contract Test 保证与本文件一致。
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface WriteCollectionMember
 */
export interface WriteCollectionMember {
    /**
     *
     * @type {WriteCollectionMemberMemberTypeEnum}
     * @memberof WriteCollectionMember
     */
    memberType: WriteCollectionMemberMemberTypeEnum;
    /**
     *
     * @type {string}
     * @memberof WriteCollectionMember
     */
    pageId?: string;
    /**
     *
     * @type {string}
     * @memberof WriteCollectionMember
     */
    entityId?: string;
    /**
     *
     * @type {string}
     * @memberof WriteCollectionMember
     */
    sortKey: string;
    /**
     *
     * @type {string}
     * @memberof WriteCollectionMember
     */
    sourceRevisionId: string;
}


/**
 * @export
 */
export const WriteCollectionMemberMemberTypeEnum = {
    Page: 'page',
    Entity: 'entity'
} as const;
export type WriteCollectionMemberMemberTypeEnum = typeof WriteCollectionMemberMemberTypeEnum[keyof typeof WriteCollectionMemberMemberTypeEnum];


/**
 * Check if a given object implements the WriteCollectionMember interface.
 */
export function instanceOfWriteCollectionMember(value: object): value is WriteCollectionMember {
    if ((!('memberType' in (value as Record<string, any>)) && !('member_type' in (value as Record<string, any>))) || ((value as Record<string, any>)['memberType'] === undefined && (value as Record<string, any>)['member_type'] === undefined)) return false;
    if ((!('sortKey' in (value as Record<string, any>)) && !('sort_key' in (value as Record<string, any>))) || ((value as Record<string, any>)['sortKey'] === undefined && (value as Record<string, any>)['sort_key'] === undefined)) return false;
    if ((!('sourceRevisionId' in (value as Record<string, any>)) && !('source_revision_id' in (value as Record<string, any>))) || ((value as Record<string, any>)['sourceRevisionId'] === undefined && (value as Record<string, any>)['source_revision_id'] === undefined)) return false;
    return true;
}

export function WriteCollectionMemberFromJSON(json: any): WriteCollectionMember {
    return WriteCollectionMemberFromJSONTyped(json, false);
}

export function WriteCollectionMemberFromJSONTyped(json: any, ignoreDiscriminator: boolean): WriteCollectionMember {
    if (json == null) {
        return json;
    }
    return {

        'memberType': json['member_type'],
        'pageId': json['page_id'] == null ? undefined : json['page_id'],
        'entityId': json['entity_id'] == null ? undefined : json['entity_id'],
        'sortKey': json['sort_key'],
        'sourceRevisionId': json['source_revision_id'],
    };
}

export function WriteCollectionMemberToJSON(json: any): WriteCollectionMember {
    return WriteCollectionMemberToJSONTyped(json, false);
}

export function WriteCollectionMemberToJSONTyped(value?: WriteCollectionMember | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'member_type': value['memberType'],
        'page_id': value['pageId'],
        'entity_id': value['entityId'],
        'sort_key': value['sortKey'],
        'source_revision_id': value['sourceRevisionId'],
    };
}